requirements.txt
fastapi
uvicorn[standard]
mangum
sqlalchemy
pydantic
pydantic-settings
bcrypt==4.0.1
passlib[bcrypt]
python-jose
psycopg2-binary
python-dotenv